Added blacklist regex for invalid characters
bb_awards was having issues where Sketch was trying to copy unicode characters that broke client side redux. So far I cannot find a good whitelist solution that would support multi-lingual so I am instead using a blacklist. If a whitelist is found in the future that works multi-lingual this could be changed. For now this fixes my issues.